rename lastmile -> logistics
