--- src/coopcycle-frontend-js/lastmile/redux.js
@@ -1,5 +1,0 @@
-// published version
-export * from 'coopcycle-frontend-js/lastmile/redux'
-
-// local version (to be used only during development)
-// export * from '../../shared-js-lib-local/lastmile/redux'
