+++ src/navigation/dispatch/Date.js
@@ -7,7 +7,7 @@ import { withTranslation } from 'react-i18next'
 
 import { connect } from 'react-redux'
 import { changeDate } from '../../redux/Dispatch/actions'
-import { selectSelectedDate } from '../../coopcycle-frontend-js/lastmile/redux'
+import { selectSelectedDate } from '../../coopcycle-frontend-js/logistics/redux'
 
 class DateScreen extends Component {
