+++ src/redux/Dispatch/actions.js
@@ -13,7 +13,7 @@ import {
   createTaskListFailure,
 
   selectSelectedDate,
-} from '../../coopcycle-frontend-js/lastmile/redux'
+} from '../../coopcycle-frontend-js/logistics/redux'
 
 import {
   startTaskSuccess,
