+++ src/redux/logistics/taskListEntityReducers.js
@@ -4,7 +4,7 @@ import {
   ASSIGN_TASK_SUCCESS,
   UNASSIGN_TASK_SUCCESS, CREATE_TASK_SUCCESS,
 } from '../Dispatch/actions';
-import { taskListUtils as utils } from '../../coopcycle-frontend-js/lastmile/redux'
+import { taskListUtils as utils } from '../../coopcycle-frontend-js/logistics/redux'
 
 const initialState = {
   byUsername: {},
